requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.2
chardet==3.0.4
discord.py==1.0.0a1735+gc4a21cc
idna==2.8
multidict==4.5.2
pycparser==2.19
PyNaCl==1.2.1
six==1.12.0
websockets==7.0
yarl==1.3.0
